Moved manageRoles to firebase folder
